✨ (main.py): introduce new functionality to allow admins to update requests for all users

+++ main.py
@@ -47,6 +47,8 @@ app.secret_key = '!secret'      # TODO: sceglierne una migliore !
 app.logger.debug("Loading configs from envs")
 app.config.from_object('config')
 
+# print(app.config)
+
 # TODO: nascondere in .env ?
 app.logger.debug("Loading OAuth configs")
 CONF_URL = 'http://127.0.0.1:8200/v1/identity/oidc/provider/default/.well-known/openid-configuration'
@@ -69,16 +71,34 @@ def homepage():
     app.logger.debug(f"/ User value: {user}")
 
     if user:
-        match request.method:
-            case 'POST':
-                dbms.add_request(user['contact']['email'])
-            case 'DELETE':
-                dbms.delete_request(user['contact']['email'])
-            case _:
-                print(request.method)
+        # ADMIN ROLE
+        if user['contact']['email'] in app.config['ADMIN_USERS']:
+            app.logger.debug(f"User {user['contact']['email']} is ADMIN")
+            user['admin'] = True
+            
+            if request.method == 'POST':
+                user_to_update = request.form['id']
+                app.logger.debug(f"UPDATING USER {user_to_update} REQUEST")
 
-        user['request_data'] = dbms.get_request_status(user['contact']['email'])
-        app.logger.debug(f"USER REQUEST DATA: {user['request_data']}")
+                dbms.update_request_status(user_to_update)
+
+            user['request_data'] = dbms.get_all_requests_status()
+            app.logger.debug(f"ADMIN REQUEST DATA: {user['request_data']}")
+
+        # USER ROLE
+        else:
+            user['admin'] = False
+
+            match request.method:
+                case 'POST':
+                    dbms.add_request(user['contact']['email'])
+                case 'DELETE':
+                    dbms.delete_request(user['contact']['email'])
+                case _:
+                    print(request.method)
+
+            user['request_data'] = dbms.get_request_status(user['contact']['email'])
+            app.logger.debug(f"USER REQUEST DATA: {user['request_data']}")
 
     return render_template('home.html', user=user)
